refactor(service): extract query params service code
to abstract params service class
